test(governance): the edit-path suite splits along the units it tests
The file had grown to 474 source lines, over the 300 the repo asks of a new
one, and it had three separable subjects sitting in it: what the builders
submit, what the form shows before anything is rendered, and what happens when
a saved edit reaches the pull lifecycle.

They are now three files, each named for its subject, and the composer state
they all start from moves into a fixture so the split does not leave three
copies of it to drift apart.

The presentation suite also picks up the report lock, including the transition
that motivated it: a source with usage history switched to cost, in both
directions, and the cases where nothing should be locked at all.
