+++ platform/app/ee/governance/dashboard/pages/__tests__/editPullSourceConfig.builders.unit.test.ts
@@ -1,10 +1,12 @@
 // SPDX-License-Identifier: LicenseRef-LangWatch-Enterprise
 
 /**
- * The edit path's half of the composer contract. Creating a source and
- * editing one differ in exactly one place — on create the upstream secret is
- * mandatory and typed in front of you; on edit it is never shown, because
- * `toDto` strips it, so a blank field has to mean "keep the stored one".
+ * What the edit form submits, and what it refuses to.
+ *
+ * Creating a source and editing one differ in exactly one place — on create
+ * the upstream secret is mandatory and typed in front of you; on edit it is
+ * never shown, because `toDto` strips it, so a blank field has to mean "keep
+ * the stored one".
  *
  * That single difference is the whole risk. `updateSource` carries the stored
  * credential across only when the key is `undefined`; a builder that helpfully
@@ -19,18 +21,13 @@
 
 import { describe, expect, it, vi } from "vitest";
 import { anthropicAdminPullConfigSchema } from "../../../services/pullers/anthropicAdmin.puller";
-import { syncIngestionPullSource } from "../../../services/pullers/ingestionPullLifecycle";
 import { recommendedPullSchedule } from "../../logic/pullCadence";
 import {
   buildAnthropicAdminPullConfig,
   buildEditedParserConfig,
   buildEditSubmission,
-  type ComposerState,
-  isBackfillStartLocked,
-  isEditablePullSource,
-  parserFieldPresentation,
-  seedComposerParserConfig,
 } from "../ingestion-sources";
+import { composer } from "./editPullSourceConfig.fixture";
 
 // `resolvePullConfig` toasts the offending field when a pull config will not
 // build. That is the behaviour under test's own reporting channel, not a
@@ -38,27 +35,6 @@ import {
 vi.mock("~/components/ui/toaster", () => ({
   toaster: { create: vi.fn() },
 }));
-
-// The lifecycle looks up the hidden governance project to address the process
-// manager. Which project it lands on is not what these tests are about.
-vi.mock("../../../services/governanceProject.service", () => ({
-  ensureHiddenGovernanceProject: vi.fn().mockResolvedValue({ id: "proj_gov" }),
-  PROJECT_KIND: { INTERNAL_GOVERNANCE: "internal_governance" },
-}));
-
-function composer(
-  parserConfig: Record<string, string>,
-  pullSchedule = "",
-): ComposerState {
-  return {
-    sourceType: "anthropic_admin",
-    name: "Anthropic org spend",
-    description: "",
-    parserConfig,
-    ottlStatements: [],
-    pullSchedule,
-  };
-}
 
 describe("buildAnthropicAdminPullConfig on the edit path", () => {
   it("given a blank secret, omits the credentials key entirely so the stored one is kept", () => {
@@ -232,29 +208,6 @@ describe("buildEditedParserConfig", () => {
   });
 });
 
-describe("isEditablePullSource", () => {
-  it("accepts a pull type the form knows how to rebuild", () => {
-    expect(isEditablePullSource("anthropic_admin")).toBe(true);
-  });
-
-  it("refuses a pull type whose blank-secret meaning is ambiguous", () => {
-    // Genie takes either a workspace token or a service-principal pair, so a
-    // blank form there could mean "keep what is stored" or "switch auth
-    // modes". Guessing locks an admin out of their own source.
-    expect(isEditablePullSource("databricks_genie")).toBe(false);
-  });
-
-  it("refuses a push type, which has no pull config to rebuild", () => {
-    expect(isEditablePullSource("otel_generic")).toBe(false);
-  });
-
-  it("refuses an absent type rather than throwing", () => {
-    // A row written by a newer deploy reaches this page as a string that misses
-    // every table; it must read as "not editable", not crash the drawer.
-    expect(isEditablePullSource(undefined)).toBe(false);
-  });
-});
-
 describe("buildEditSubmission", () => {
   const pullSource = {
     id: "src_1",
@@ -390,238 +343,3 @@ describe("buildEditSubmission", () => {
  * and the lifecycle correctly read `null` as "disable". Only a test that runs
  * the edit path's output through the thing that consumes it can fail.
  */
-describe("a saved edit reaching the pull lifecycle", () => {
-  function sourceRowFrom(submission: { pullSchedule?: string | null }) {
-    return {
-      id: "src_1",
-      organizationId: "org_1",
-      status: "active",
-      archivedAt: null,
-      updatedAt: new Date("2026-01-01T00:00:00Z"),
-      pollerCursor: null,
-      pullSchedule: submission.pullSchedule ?? null,
-    } as unknown as Parameters<typeof syncIngestionPullSource>[0]["source"];
-  }
-
-  async function syncAfterEdit(pullSchedule: string) {
-    const submission = buildEditSubmission({
-      organizationId: "org_1",
-      source: {
-        id: "src_1",
-        sourceType: "anthropic_admin",
-        parserConfig: {
-          adapter: "anthropic_admin",
-          report: "usage",
-          bucketWidth: "1h",
-          schedule: "0 * * * *",
-        },
-      },
-      name: "Anthropic org spend",
-      description: "",
-      parserConfig: {
-        credentialsToken: "",
-        report: "usage",
-        bucketWidth: "1h",
-      },
-      ottlStatements: [],
-      pullSchedule,
-    });
-    const commands = { configure: vi.fn(), disable: vi.fn() };
-
-    await syncIngestionPullSource({
-      prisma: {} as never,
-      source: sourceRowFrom(submission ?? {}),
-      commands,
-    });
-
-    return { submission, commands };
-  }
-
-  it("keeps pulling after an admin clears the cadence field", async () => {
-    const { submission, commands } = await syncAfterEdit("   ");
-
-    // The lifecycle outcome first: this is the assertion that fails when the
-    // edit path goes back to saving null, and the one that says what the
-    // admin actually loses — a source that has stopped pulling.
-    expect(commands.disable).not.toHaveBeenCalled();
-    expect(commands.configure).toHaveBeenCalledWith(
-      expect.objectContaining({
-        sourceId: "src_1",
-        cron: recommendedPullSchedule("anthropic_admin"),
-      }),
-    );
-    expect(submission?.pullSchedule).toBe(
-      recommendedPullSchedule("anthropic_admin"),
-    );
-  });
-
-  it("still runs on an explicitly typed cadence", async () => {
-    const { commands } = await syncAfterEdit("0 */2 * * *");
-
-    expect(commands.disable).not.toHaveBeenCalled();
-    expect(commands.configure).toHaveBeenCalledWith(
-      expect.objectContaining({ cron: "0 */2 * * *" }),
-    );
-  });
-});
-
-describe("isBackfillStartLocked", () => {
-  it("locks the start for a usage source that has already pulled", () => {
-    // The usage cursor never rewinds, so an edit here would silently do
-    // nothing.
-    expect(isBackfillStartLocked({ hasPulled: true, report: "usage" })).toBe(
-      true,
-    );
-  });
-
-  it("leaves the start editable for a cost source that has already pulled", () => {
-    // The cost cursor binds `startingAt` into its identity: moving the start
-    // discards the cursor and re-reads the widened window. That is the repair
-    // lever for wrong early figures, and locking it would remove it.
-    expect(isBackfillStartLocked({ hasPulled: true, report: "cost" })).toBe(
-      false,
-    );
-  });
-
-  it("leaves the start editable before the first pull, whatever the report", () => {
-    expect(isBackfillStartLocked({ hasPulled: false, report: "usage" })).toBe(
-      false,
-    );
-    expect(isBackfillStartLocked({ hasPulled: false, report: "cost" })).toBe(
-      false,
-    );
-  });
-
-  it("does not lock a source whose config carries no report", () => {
-    // Claiming immutability we cannot justify is the worse error: it sends an
-    // admin to archive-and-recreate over a field that may well be editable.
-    expect(isBackfillStartLocked({ hasPulled: true, report: undefined })).toBe(
-      false,
-    );
-  });
-});
-
-describe("parserFieldPresentation", () => {
-  const secretField = {
-    key: "credentialsToken",
-    label: "Admin API key",
-    placeholder: "sk-ant-admin-...",
-    hint: "Generate one in the Anthropic console.",
-    required: true,
-  };
-  const plainField = {
-    key: "report",
-    label: "Report",
-    placeholder: "usage",
-    hint: "usage or cost",
-    required: true,
-  };
-
-  it("on edit, tells the admin a blank secret keeps the current key", () => {
-    const p = parserFieldPresentation({ field: secretField, mode: "edit" });
-
-    expect(p.hint).toContain("Leave blank to keep the current key");
-    expect(p.placeholder).toBe("Unchanged");
-  });
-
-  it("on edit, drops the required marker from a secret", () => {
-    // A field that says "leave blank to keep the current key" while still
-    // carrying a required marker is a form contradicting itself.
-    expect(
-      parserFieldPresentation({ field: secretField, mode: "edit" }).isRequired,
-    ).toBe(false);
-  });
-
-  it("on create, keeps the secret mandatory and its own hint", () => {
-    const p = parserFieldPresentation({ field: secretField, mode: "create" });
-
-    expect(p.isRequired).toBe(true);
-    expect(p.hint).toBe(secretField.hint);
-    expect(p.placeholder).toBe(secretField.placeholder);
-  });
-
-  it("leaves a non-secret field identical in both modes", () => {
-    expect(
-      parserFieldPresentation({ field: plainField, mode: "edit" }),
-    ).toEqual(parserFieldPresentation({ field: plainField, mode: "create" }));
-  });
-
-  it("marks the DSL fields as multiline", () => {
-    for (const key of ["parserDsl", "eventMappingDsl"]) {
-      expect(
-        parserFieldPresentation({
-          field: { ...plainField, key },
-          mode: "create",
-        }).isMultiline,
-      ).toBe(true);
-    }
-    expect(
-      parserFieldPresentation({ field: plainField, mode: "create" })
-        .isMultiline,
-    ).toBe(false);
-  });
-});
-
-describe("seedComposerParserConfig", () => {
-  it("maps a stored config back onto the form fields", () => {
-    const values = seedComposerParserConfig({
-      sourceType: "anthropic_admin",
-      storedParserConfig: {
-        adapter: "anthropic_admin",
-        report: "usage",
-        bucketWidth: "1h",
-        startingAt: "2026-08-01T00:00:00.000Z",
-      },
-    });
-
-    expect(values.report).toBe("usage");
-    expect(values.bucketWidth).toBe("1h");
-    expect(values.startingAt).toBe("2026-08-01T00:00:00.000Z");
-  });
-
-  it("leaves every secret field blank rather than pre-filling it", () => {
-    const values = seedComposerParserConfig({
-      sourceType: "anthropic_admin",
-      storedParserConfig: { report: "usage" },
-    });
-
-    expect(values.credentialsToken ?? "").toBe("");
-  });
-
-  it("never surfaces a stored credential into the form, even if one reaches it", () => {
-    // `toDto` strips `credentials` today, so this should be unreachable. It is
-    // asserted anyway because the cost of that stripping regressing is the
-    // form posting an envelope straight back into the server's replay refusal.
-    const values = seedComposerParserConfig({
-      sourceType: "anthropic_admin",
-      storedParserConfig: {
-        report: "usage",
-        credentials: "enc:v1:deadbeef:cafe:0123",
-      },
-    });
-
-    expect(Object.values(values).join(" ")).not.toContain("enc:v1:");
-    expect(values.credentialsToken ?? "").toBe("");
-  });
-
-  it("round-trips an untouched form without proposing any credential change", () => {
-    const stored = {
-      adapter: "anthropic_admin",
-      report: "usage",
-      bucketWidth: "1h",
-      schedule: "0 * * * *",
-    };
-    const values = seedComposerParserConfig({
-      sourceType: "anthropic_admin",
-      storedParserConfig: stored,
-    });
-
-    const config = buildAnthropicAdminPullConfig(
-      composer(values, "0 * * * *"),
-      { requireCredentials: false },
-    );
-
-    expect(config).not.toHaveProperty("credentials");
-    expect(anthropicAdminPullConfigSchema.parse(config).report).toBe("usage");
-  });
-});
